Rename variant.h to OverloadSet.h
[why]
The header is no longer related to std::variant in any way. It only
contains the OverloadSet type.
