+++ include/gul17/meson.build
@@ -11,6 +11,7 @@ standalone_headers = [
     'hexdump.h',
     'join_split.h',
     'num_util.h',
+    'OverloadSet.h',
     'replace.h',
     'SlidingBuffer.h',
     'SmallVector.h',
@@ -26,7 +27,6 @@ standalone_headers = [
     'Trigger.h',
     'trim.h',
     'type_name.h',
-    'variant.h',
 ]
 
 libgul_sub_headers = files(standalone_headers)
